requirements.txt
google-generativeai
google-adk
pydantic
python-dotenv
asyncio